rename: 변수 위치 이동, 이미지 파일명 변경
