+++ client/src/components/login/SocialLogin/index.tsx
@@ -1,7 +1,7 @@
 import SocialLoginButton from '../SocialLoginButton';
-import naver_circle from 'assets/images/naver_circle.svg';
-import kakao_circle from 'assets/images/kakao_circle.svg';
-import google_circle from 'assets/images/google_circle.svg';
+import socialLogin_naver from 'assets/images/socialLogin_naver.svg';
+import socialLogin_kakao from 'assets/images/socialLogin_kakao.svg';
+import socialLogin_google from 'assets/images/socialLogin_google.svg';
 import styles from './SocialLogin.module.scss';
 import classNames from 'classnames/bind';
 
@@ -15,16 +15,24 @@ const SocialLogin = () => (
     <ul className={cn('providers')}>
       {/* TODO */}
       <li>
-        <SocialLoginButton href="#" provider="네이버" imgSRC={naver_circle} />
+        <SocialLoginButton
+          href="#"
+          provider="네이버"
+          imgSRC={socialLogin_naver}
+        />
       </li>
       <li>
-        <SocialLoginButton href="#" provider="카카오" imgSRC={kakao_circle} />
+        <SocialLoginButton
+          href="#"
+          provider="카카오"
+          imgSRC={socialLogin_kakao}
+        />
       </li>
       <li>
         <SocialLoginButton
           href="http://localhost:9001/oauth2/authorization/google"
           provider="구글"
-          imgSRC={google_circle}
+          imgSRC={socialLogin_google}
         />
       </li>
     </ul>
