+++ client/src/components/signup/SignupDetailForm/index.tsx
@@ -99,7 +99,6 @@ const SignupDetailForm = () => {
       setSignupFormData(data => ({ ...data, nickname }));
       increaseStep();
     } catch (e) {
-      console.log(e);
       showToastMessage('회원가입 중 문제가 발생했습니다');
     }
   };
