+++ client/src/types/auth.ts
@@ -11,3 +11,15 @@ export interface SignupDetailFormData {
   nickname: string;
 }
 export type SignupFormData = SignupBasicFormData & SignupDetailFormData;
+
+export interface NewUser {
+  email: string;
+  password: string;
+  nickname: string;
+  gender: string;
+  birthYear: number;
+  bicycleType: string;
+  bicycleCareer: number;
+  introduce: string | null;
+  isTermsEnable: boolean;
+}
